Move NoiseLearnerV3 decoder
