+++ qiskit_ibm_runtime/decoders/noise_learner_v3/__init__.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2025-2026.
+# (C) Copyright IBM 2026.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -10,15 +10,4 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Converters for NLV3."""
-
-from .version_0_1 import (
-    noise_learner_v3_inputs_from_0_1,
-    noise_learner_v3_inputs_to_0_1,
-    noise_learner_v3_result_to_0_1,
-)
-from .version_0_2 import (
-    noise_learner_v3_inputs_from_0_2,
-    noise_learner_v3_inputs_to_0_2,
-    noise_learner_v3_result_to_0_2,
-)
+"""Result decoder for NoiseLearnerV3."""
